requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.22
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
